requirements.txt
Flask==0.12.2
Flask_Script==2.0.5
pytest==3.1.3
setuptools==36.2.0
Flask_Bcrypt==0.7.1
Flask_Testing==0.6.2
flask_mongoengine==0.9.3
itsdangerous==0.24
